Consolidate tweet updating
Reconcile differences between the two ways that tweets were being added to
the html and put the result into a new function.

One of the previous methods to add tweets was to start from the end of
the array and move to the front, appending new elements to the end of
the body element. However, while this method is good for the initial
set of tweets, it is not easily adaptable to add new tweets of the
array grows. Thus, it was switched for a method that starts in the
front of the array and counts up, appending new elements to the top of
the previous element.