Adjust tweet display implementation
Initially, when a user link is pressed, all tweets from a specific
user would be displayed even though they were not yet updated on the
home tweet feed because the update button was not pressed yet. This
was changed so that when the user link is pressed, only tweets
that are already displayed on the home feed are filtered.